Added eject function
