+++ src/client/client.js
@@ -1,3 +1,4 @@
+// @remove-file-on-eject
 import React from 'react'
 import { BrowserRouter as Router } from 'react-router-dom'
 import ReactDOM from 'react-dom'
