--- src/client/index.js
@@ -1,3 +1,0 @@
-import client from './client'
-
-export default client
