+++ src/config/default.client.js
@@ -1,4 +1,3 @@
-// @remove-file-on-eject
 // ========================================================
 // Default Configuration For Client
 // ========================================================
