+++ src/server/index.js.eject
@@ -1,0 +1,3 @@
+import server from './server'
+
+export default server
